Refactor to external JS
