+++ .gitignore
@@ -8,3 +8,4 @@ coverage
 phpunit.xml
 psalm.xml
 vendor
+/node_modules
