+++ package.json
@@ -1,0 +1,25 @@
+{
+    "main": "dist/form-components.js",
+    "name": "laravel-form-components",
+    "scripts": {
+        "build": "npx rollup -c",
+        "watch": "npx rollup -c -w"
+    },
+    "author": "Randall Wilk",
+    "license": "MIT",
+    "devDependencies": {
+        "@babel/core": "^7.12.3",
+        "@babel/plugin-proposal-object-rest-spread": "^7.12.1",
+        "@babel/preset-env": "^7.12.1",
+        "@rollup/plugin-alias": "^3.1.1",
+        "@rollup/plugin-babel": "^5.2.1",
+        "@rollup/plugin-commonjs": "^15.1.0",
+        "@rollup/plugin-node-resolve": "^9.0.0",
+        "fs-extra": "^9.0.1",
+        "md5": "^2.3.0",
+        "rollup": "^2.32.1",
+        "rollup-plugin-filesize": "^9.0.2",
+        "rollup-plugin-output-manifest": "^1.0.2",
+        "rollup-plugin-terser": "^7.0.2"
+    }
+}
