+++ resources/js/components/index.js
@@ -1,0 +1,3 @@
+import customSelect from './custom-select';
+
+window.customSelect = customSelect;
