+++ src/Controllers/Concerns/CanPretendToBeAFile.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace Rawilk\FormComponents\Controllers\Concerns;
+
+/**
+ * This concern is an exact copy of the one used by Laravel Livewire...
+ */
+trait CanPretendToBeAFile
+{
+    public function pretendResponseIsFile($file)
+    {
+        $expires = strtotime('+1 year');
+        $lastModified = filemtime($file);
+        $cacheControl = 'public, max-age=31536000';
+
+        if ($this->matchesCache($lastModified)) {
+            return response()->make('', 304, [
+                'Expires' => $this->httpDate($expires),
+                'Cache-Control' => $cacheControl,
+            ]);
+        }
+
+        return response()->file($file, [
+            'Content-Type' => 'application/javascript; charset=utf-8',
+            'Expires' => $this->httpDate($expires),
+            'Cache-Control' => $cacheControl,
+            'Last-Modified' => $this->httpDate($lastModified),
+        ]);
+    }
+
+    protected function matchesCache($lastModified): bool
+    {
+        $ifModifiedSince = $_SERVER['HTTP_IF_MODIFIED_SINCE'] ?? '';
+
+        return @strtotime($ifModifiedSince) === $lastModified;
+    }
+
+    protected function httpDate($timestamp): string
+    {
+        return sprintf('%s GMT', gmdate('D, d M Y H:i:s', $timestamp));
+    }
+}
