+++ src/FormComponents.php
@@ -6,57 +6,93 @@ namespace Rawilk\FormComponents;
 
 final class FormComponents
 {
-    private static array $styles = [];
-    private static array $scripts = [];
+    private array $styles = [];
+    private array $scripts = [];
 
-    public static function addStyle(string $style): void
+    public function addStyle(string $style): void
     {
-        if (! in_array($style, static::$styles, true)) {
-            static::$styles[] = $style;
+        if (! in_array($style, $this->styles, true)) {
+            $this->styles[] = $style;
         }
     }
 
-    public static function styles(): array
+    public function styles(): array
     {
-        return static::$styles;
+        return $this->styles;
     }
 
-    public static function outputStyles(bool $force = false): string
+    public function outputStyles(bool $force = false): string
     {
-        if (! $force && static::scriptsAreDisabled()) {
+        if (! $force && $this->scriptsAreDisabled()) {
             return '';
         }
 
-        return collect(static::$styles)->map(function (string $style) {
+        return collect($this->styles)->map(function (string $style) {
             return '<link rel="stylesheet" href="' . $style . '" />';
         })->implode(PHP_EOL);
     }
 
-    public static function addScript(string $script): void
+    public function addScript(string $script): void
     {
-        if (! in_array($script, static::$scripts, true)) {
-            static::$scripts[] = $script;
+        if (! in_array($script, $this->scripts, true)) {
+            $this->scripts[] = $script;
         }
     }
 
-    public static function scripts(): array
+    public function scripts(): array
     {
-        return static::$scripts;
+        return $this->scripts;
     }
 
-    public static function outputScripts(bool $force = false): string
+    public function outputScripts(bool $force = false, $options = []): string
     {
-        if (! $force && static::scriptsAreDisabled()) {
-            return '';
+        $js = $this->javaScript($options);
+
+        if (! $force && $this->scriptsAreDisabled()) {
+            return $js;
         }
 
-        return collect(static::$scripts)->map(function (string $script) {
+        return $js . collect($this->scripts)->map(function (string $script) {
             return '<script src="' . $script . '"></script>';
         })->implode(PHP_EOL);
     }
 
-    private static function scriptsAreDisabled(): bool
+    /**
+     * This will output the JavaScript necessary to run some components
+     * such as CustomSelect.
+     *
+     * @param array $options
+     * @return string
+     */
+    public function javaScript(array $options = []): string
     {
+        $html = config('app.debug') ? ['<!-- FormComponents Scripts -->'] : [];
+
+        $html[] = $this->javaScriptAssets($options);
+
+        return implode(PHP_EOL, $html);
+    }
+
+    private function scriptsAreDisabled(): bool
+    {
+        if (config('form-components.link_vendor_cdn_assets') === false) {
+            return true;
+        }
+
         return ! config('app.debug');
+    }
+
+    private function javaScriptAssets(array $options = []): string
+    {
+        $appUrl = config('form-components.asset_url', rtrim($options['asset_url'] ?? '', '/'));
+
+        $manifest = json_decode(file_get_contents(__DIR__ . '/../dist/manifest.json'), true);
+        $versionedFileName = $manifest['/form-components.js'];
+
+        $fullAssetPath = "{$appUrl}/form-components{$versionedFileName}";
+
+        return <<<HTML
+        <script src="{$fullAssetPath}" data-turbolinks-eval="false"></script>
+        HTML;
     }
 }
